WF1 fix

+++ sources/util/CustomMetrics.py
@@ -41,6 +41,6 @@ class WF1(tfa.metrics.F1Score):
         super(WF1, self).__init__(num_classes, average="weighted", name="WF1", **kwargs)
 
     def update_state(self, y_true, y_pred, sample_weight=None):
-        y_true = tf.expand_dims(tf.argmax(y_true, axis=-1), axis=-1)
-        y_pred = tf.expand_dims(tf.argmax(y_pred, axis=-1), axis=-1)
+        y_true = tf.cast(tf.expand_dims(tf.argmax(y_true, axis=-1), axis=-1), dtype=tf.float32)
+        y_pred = tf.cast(tf.expand_dims(tf.argmax(y_pred, axis=-1), axis=-1), dtype=tf.float32)
         super(WF1, self).update_state(y_true, y_pred, sample_weight)
